Network: Rename client to tcpClient.
